requirements.txt
flask==2.2.3
flask-sqlalchemy==3.0.3
flask-cors==3.0.10
# psycopg2-binary==2.9.9
python-dotenv==1.0.0
werkzeug==2.2.3